refactor(types): types 파일 정리
