+++ src/components/auth/SocialLoginButton.tsx
@@ -1,4 +1,4 @@
-import {SocialLoginButtonProps} from "@/types/featureProps";
+import {SocialLoginButtonProps} from "@/types/components/featureProps";
 import {useSocialAuth} from "@/lib/hooks/useSocialAuth";
 import {Button} from "@/components/common/Button";
 import Image from "next/image";
