+++ src/components/explore/MentorCard.tsx
@@ -1,4 +1,4 @@
-import {MentorCardProps} from "@/types/featureProps";
+import {MentorCardProps} from "@/types/components/featureProps";
 import {Card} from "@/components/common/Card";
 import Image from "next/image";
 import {formatDeveloperTitle, getTechDisplayName} from "@/lib/utils/category";
