+++ src/lib/utils/category.ts
@@ -1,6 +1,6 @@
-import {FilterKey, FilterOption, TechCategories, TechOption} from "@/types/category";
+import {FilterKey, FilterOption, TechCategories, TechOption} from "@/types/shared/category";
 import {experienceOptions, ratingOptions, TECH_CATEGORIES} from "@/config/category";
-import {DeveloperInterest, DeveloperLevel} from "@/types/mentor";
+import {MentorInterest, MentorLevel} from "@/types/core/mentor";
 
 export const getAllTechOptions = (categories: TechCategories): TechOption[] => {
     return Object.values(categories).flat();
@@ -25,14 +25,14 @@ export const getTechDisplayName = (value: string): string => {
     return tech?.label || value;
 };
 
-export const calculateDeveloperLevel = (experience: number): DeveloperLevel => {
+export const calculateDeveloperLevel = (experience: number): MentorLevel => {
     if (experience >= 8) return 'Lead';
     if (experience >= 5) return 'Senior';
     if (experience >= 3) return 'Mid';
     return 'Junior';
 };
 
-export const formatDeveloperTitle = (experience: number, interest: DeveloperInterest): string => {
+export const formatDeveloperTitle = (experience: number, interest: MentorInterest): string => {
     const level = calculateDeveloperLevel(experience);
     return `${level} ${interest} Developer`;
 };
