+++ src/types/shared/params.ts
@@ -1,0 +1,10 @@
+export interface SearchParams {
+    search?: string;
+    techStack?: string[];
+    experience?: string[];
+    rating?: string[];
+    priceMin?: number;
+    priceMax?: number;
+    page?: number;
+    limit?: number;
+}
